clang-tidy tweaks #3300

+++ cpp/core/dwi/tractography/ACT/gmwmi.h
@@ -39,12 +39,14 @@ protected:
 
 public:
   GMWMI_finder(const Image<float> &buffer)
-      : interp_template(buffer), min_vox(std::min(buffer.spacing(0), std::min(buffer.spacing(1), buffer.spacing(2)))) {
+      : interp_template(buffer),
+        min_vox(static_cast<float>(std::min(buffer.spacing(0), std::min(buffer.spacing(1), buffer.spacing(2))))) {
     ACT::debug_validate_5TT_image(buffer);
   }
 
   GMWMI_finder(const Interp &interp)
-      : interp_template(interp), min_vox(std::min(interp.spacing(0), std::min(interp.spacing(1), interp.spacing(2)))) {
+      : interp_template(interp),
+        min_vox(static_cast<float>(std::min(interp.spacing(0), std::min(interp.spacing(1), interp.spacing(2))))) {
     ACT::debug_validate_5TT_image(interp);
   }
